requirements.txt
requests
python-dotenv
feedparser
sentence-transformers
anthropic
numpy
